requirements.txt
bitsandbytes
datasets
faster-whisper
librosa
mutagen
numpy
openai-whisper
pydub
soundfile
torch
torchaudio
transformers==4.48.1
openai
python-dotenv
jiwer
pandas
peft
evaluate
webvtt-py
scipy
tqdm
pydantic
accelerate
boto3
google-cloud-speech
google-cloud-storage
amazon-transcribe
hebrew
runpod
